+++ CHANGELOG.md
@@ -14,7 +14,14 @@
 - `Arc.Fillet`
 - `Ellipse`
 - `EllipticalArc`
+- `IndexedPolycurve`
 - `Grid1d.GetCellDomains`
+- `Message.Info`
+- `Message.Error`
+- `Message.Warning`
+- `Topography.Trimmed`
+- `new Topography(Topography other)`
+- `Topography.TopMesh()`
 
 ### Changed
 
@@ -34,6 +41,7 @@
 - `EdgeInfo`: obsolete attribute is removed from `HasVerticalChange` property.
 - `RoutingConfiguration`: removed obsolete `MainLayer` and `LayerPenalty` properties.
 - `Material.EmissiveFactor` is now 0.0 by default.
+- `Polygon.Area()` now returns an unsigned area by default, accepts a bool `signed` parameter to preserve the previous signed behavior.
 
 ### Fixed
 
@@ -44,6 +52,8 @@
 - `Vector3.AreCoplanar` would sometimes return false negatives.
 - Fix the polygon centroid calculation to remove collinear vertices.
 - Fix the tests for 3dCentroid testing.
+- `Message` created from `Message.FromPoint` now has `Transform.Origin` set exactly on original point.
+- Certain `Triangle` constructors would not correctly update `Vertex.Triangles`, this is fixed.
 
 ## 1.6.0